requirements.txt
appdirs==1.4.0
coreapi==2.2.4
coreschema==0.0.4
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.10
django-braces==1.11.0
django-cors-headers==2.0.2
django-crispy-forms==1.6.1
django-filter==1.0.1
django-oauth-toolkit==0.11.0
django-rest-swagger==2.1.1
django-toolbelt==0.0.1
djangorestframework==3.5.4
gunicorn==19.6.0
itypes==1.1.0
Jinja2==2.9.5
MarkupSafe==0.23
oauthlib==1.1.2
olefile==0.44
openapi-codec==1.3.1
packaging==16.8
Pillow==4.0.0
psycopg2==2.6.2
pyparsing==2.1.10
requests==2.13.0
simplejson==3.10.0
six==1.10.0
static3==0.7.0
uritemplate==3.0.0
whitenoise==3.3.0
